[ progressive ] RefC runtime: fix uninitialized constructor args + disable unique reuse
Three runtime fixes for the RefC backend:

1. Zero-initialize constructor args in idris2_newConstructor. The codegen
   emits idris2_removeReference on all args even for freshly allocated
   constructors (not just reused ones). Without memset, the cleanup loop
   dereferences garbage pointers — a classic Heisenbug that only crashes
   without a debugger (debug heap fills with 0xCD which has odd low bits
   → tagged as unboxed → removeReference returns harmlessly).

2. Disable idris2_isUnique optimization. The unique reuse path has a
   double-free bug: old args are freed then reassigned, causing
   use-after-free in reverseOnto, mapAppend, and similar list operations.
   Proper fix requires codegen changes in RefC.idr.

3. Widen refCounter from uint16_t to uint32_t. The old 65535 ceiling
   caused premature immortalization during long compilations (296 modules),
   leaking all widely-shared closures and strings.

Also adds pool_flush_all for malloc failure recovery, RSS logging,
and improved memory statistics infrastructure.

+++ support/refc/_datatypes.h
@@ -35,10 +35,16 @@ typedef struct {
   // Objects that reach the maximum reference count will be immortalized.
   // This 'immortalization' feature is also utilized to prevent statically
   // allocated objects from being destroyed.
-#define IDRIS2_VP_REFCOUNTER_MAX UINT16_MAX
-  uint16_t refCounter;
+  //
+  // WIDENED from uint16_t to uint32_t: the old 65535 ceiling caused
+  // premature immortalization during long compilations (296 modules),
+  // leaking all widely-shared closures and strings.  On 64-bit targets
+  // the struct stays 8 bytes thanks to existing alignment padding.
+#define IDRIS2_VP_REFCOUNTER_MAX UINT32_MAX
+  uint32_t refCounter;
   uint8_t tag;
   uint8_t reserved;
+  // 2 bytes padding on 64-bit (absorbed from old post-header padding)
 } Value_header;
 #define IDRIS2_STOCKVAL(t)                                                     \
   { IDRIS2_VP_REFCOUNTER_MAX, t, 0 }
